pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memo-transcriber"
version = "0.1.0"
description = "Voice memo transcription tool"
requires-python = ">=3.13"
dependencies = ["click","tqdm", "openai-whisper", "faster-whisper", "pyobjc-framework-Speech", "pyobjc-framework-AVFoundation"]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "types-click>=7.1.0",
]

# Define console commands
[project.scripts]
memo-transcriber = "memo_transcriber.cli:main"

# Hatchling configuration
[tool.hatch.build.targets.wheel]
packages = ["src/memo_transcriber"]
# Exclude development files from production builds
exclude = [
    "src/memo_transcriber/dev_*.py",
    "test_*.db",
    "*.db-journal",
    ".dev"
]

[tool.hatch.build.targets.sdist]
# Also exclude from source distribution
exclude = [
    "src/memo_transcriber/dev_*.py",
    "test_*.db",
    "*.db-journal",
    ".dev"
]
